// src/core/models/groups.ts
export type GroupType = {
  id: string;
  createdAt: string;
  updatedAt: string;
  version: number;
  name: string;
  code: string;
  enterprise: string;
  enterpriseContacts: EnterpriseContactResponseType[];
  archived: boolean;
  deleted: boolean;
}
// ...
export type CreateGroupType = {
  name: string,
  code: string,
  enterpriseId: string
}